+++ heksher/clients/subclasses.py
@@ -15,8 +15,9 @@ from weakref import WeakValueDictionary
 from httpx import Response
 from ordered_set import OrderedSet
 
+import heksher.main_client
 from heksher.clients.util import collate_rules
-from heksher.heksher_client import BaseHeksherClient
+from heksher.heksher_client import BaseHeksherClient, TemporaryClient
 from heksher.setting import RuleBranch, Setting
 
 logger = getLogger(__name__)
@@ -59,7 +60,7 @@ class V1APIClient(BaseHeksherClient, ABC):
         """
         return collate_rules(self._context_features, rules)
 
-    def add_settings(self, settings):
+    def add_settings(self, settings: Iterable[Setting]):
         for s in settings:
             self._undeclared.put_nowait(s)
 
@@ -140,6 +141,39 @@ class V1APIClient(BaseHeksherClient, ABC):
                 logger.warning('context feature value is not tracked by client',
                                extra={'context_feature': k, 'context_feature_value': v})
         return super().context_namespace(user_namespace)
+
+    def _set_as_main(self):
+        """
+        Transfer "main-ness" to self, either from a temporary client or another v1 client
+        """
+        previous_main: BaseHeksherClient = heksher.main_client.Main  # for readability of this function
+        if isinstance(previous_main, TemporaryClient):
+            self.add_settings(previous_main.undeclared)
+        elif isinstance(previous_main, V1APIClient):
+            # know when you've fucked up
+            logger.warning("switching main heksher clients! this is NOT recommended! (unless this is a test)")
+            if not previous_main._undeclared.empty():
+                raise RuntimeError("previous main heksher client still has unprocessed declarations, "
+                                   "did you forget to close it?")
+            if self._context_features != previous_main._context_features:
+                # now you've really fucked up
+                raise RuntimeError(f"new main heksher client has different contexts, "
+                                   f"previous: {previous_main._context_features}, new: {self._context_features}")
+            # if we are using the same contexts, we can safely add the same settings to this client
+            # no need to redeclare them, so we just track them
+            for setting_name, setting in previous_main._tracked_settings.items():
+                if setting_name in self._tracked_settings:
+                    continue
+                self._tracked_settings[setting.name] = setting
+            if self._tracked_context_options != previous_main._tracked_context_options:
+                # this won't cause errors, but it will cause some settings to have the wrong values, be warned
+                logger.warning("new main heksher client tracks different context options",
+                               extra={"previous_options": previous_main._tracked_context_options,
+                                      "new_options": self._tracked_context_options})
+        else:
+            raise TypeError(f'cannot change main client from type '
+                            f'{type(previous_main).__name__} to type {type(self).__name__}')
+        heksher.main_client.Main = self
 
 
 class ContextFeaturesMixin(BaseHeksherClient, ABC):