+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ### Added
 * Added add_validator method for Setting; Validators added to each Setting will be called by order they were added 
   (each validator will receive value of previous validator); Validators will be called each time the Setting is updated
-### 0.1.4
+* Added close checks to async client exit
+### Changed
+* Changing main client between similar clients that track the same contexts is now possible, but not recommended. Changed for testing purposes.
+## 0.1.4
 ### Changed
 * Changed httpx version limitation in poetry.toml to `*`
 ## 0.1.3